+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added a `last_used_at` field to API token.
+- Added an upload progress bar to the CLI.
 
 ## [0.1.0] - 2026-06-13